+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Show settings menu item in titlebar menu
 - Sidebar closes on escape key
+- add chat name to webxdc window title
 
 ### Fixed
 - Fix crash on migrating accounts from an older version (before 1.21.0)